cloud_functions/meta_daily_sync/requirements.txt
# Cloud Function dependencies
functions-framework==3.*
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.16.4
requests==2.31.0
pyyaml==6.0.1
tenacity==8.2.3
python-dateutil==2.8.2
